Move tests into separate packages
